requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.3.0
black==19.10b0
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
Click==7.0
command-not-found==0.3
cryptography==2.1.4
defer==1.0.6
devscripts===2.17.12ubuntu1.1
discord.py==1.3.1
evdev==0.7.0
gpg==1.10.0
httplib2==0.9.2
idna==2.8
idna-ssl==1.1.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
louis==3.5.0
Mako==1.0.7
MarkupSafe==1.0
meson==0.49.2
multidict==4.7.4
netifaces==0.10.4
olefile==0.45.1
onboard==1.4.1
pathspec==0.7.0
pexpect==4.2.1
Pillow==5.1.0
psycopg2-binary==2.8.4
pycairo==1.16.2
pycrypto==2.6.1
pygobject==3.26.1
python-apt===1.6.5-ubuntu0.2-elementary14-ubuntu5.1.1
python-debian==0.1.32
python-dotenv==0.10.5
python-magic==0.4.16
pyudev==0.21.0
pyxdg==0.25
PyYAML==3.12
regex==2020.2.20
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
six==1.11.0
toml==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unidiff==0.5.4
urllib3==1.22
virtualenv==15.1.0
websockets==8.1
xkit==0.0.0
yarl==1.4.2
